+++ spec/rubocop/cop/style/guard_clause_spec.rb
@@ -132,7 +132,15 @@ RSpec.describe RuboCop::Cop::Style::GuardClause do
       end
 
       def func
-        work if something
+        work unless something
+      end
+    RUBY
+  end
+
+  it 'accepts a method with empty parentheses as its body' do
+    expect_no_offenses(<<~RUBY)
+      def func
+        ()
       end
     RUBY
   end